requirements.txt
# Core Framework
pytest==7.4.4
pytest-cov==4.1.0
pytest-xdist==3.5.0
allure-pytest==2.13.2

# HTTP & Async
requests==2.31.0
httpx==0.26.0

# Data Handling
PyYAML==6.0.1
pydantic==2.5.3
pandas==2.1.4
openpyxl==3.1.2
jsonschema==4.21.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.25

# Data Generation
Faker==22.2.0

# Utilities
python-dotenv==1.0.0
cachetools==5.3.2
loguru==0.7.2
deepdiff==6.7.1
